feat(715): moove all additional skills progress to student progress feature
