+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/infrastructure/adapter/service/AdditionalSkillServiceConfig.java
@@ -1,10 +1,7 @@
 package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.service;
 
-import fr.avenirsesr.portfolio.additionalskill.domain.port.input.AdditionalSkillService;
 import fr.avenirsesr.portfolio.additionalskill.domain.port.output.OpenSearchIndex;
 import fr.avenirsesr.portfolio.additionalskill.domain.port.output.repository.AdditionalSkillRepository;
-import fr.avenirsesr.portfolio.additionalskill.domain.service.AdditionalSkillServiceImpl;
-import fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.repository.AdditionalSkillProgressDatabaseRepository;
 import fr.avenirsesr.portfolio.interoperability.additionalskill.rome.domain.port.input.RomeAdditionalSkillService;
 import fr.avenirsesr.portfolio.interoperability.additionalskill.rome.domain.port.output.RomeAdditionalSkillApi;
 import fr.avenirsesr.portfolio.interoperability.additionalskill.rome.domain.port.output.repository.Rome4VersionRepository;
@@ -16,29 +13,20 @@ import org.springframework.context.annotation.Configuration;
 @Slf4j
 @Configuration
 public class AdditionalSkillServiceConfig {
-  private final AdditionalSkillProgressDatabaseRepository additionalSkillProgressRepository;
   private final AdditionalSkillRepository additionalSkillRepository;
   private final Rome4VersionRepository rome4VersionRepository;
   private final RomeAdditionalSkillApi romeAdditionalSkillApi;
   private final OpenSearchIndex openSearchIndex;
 
   public AdditionalSkillServiceConfig(
-      AdditionalSkillProgressDatabaseRepository additionalSkillProgressRepository,
       AdditionalSkillRepository additionalSkillRepository,
       Rome4VersionRepository rome4VersionRepository,
       RomeAdditionalSkillApi romeAdditionalSkillApi,
       OpenSearchIndex openSearchIndex) {
-    this.additionalSkillProgressRepository = additionalSkillProgressRepository;
     this.additionalSkillRepository = additionalSkillRepository;
     this.rome4VersionRepository = rome4VersionRepository;
     this.romeAdditionalSkillApi = romeAdditionalSkillApi;
     this.openSearchIndex = openSearchIndex;
-  }
-
-  @Bean
-  public AdditionalSkillService additionalSkillService() {
-    return new AdditionalSkillServiceImpl(
-        additionalSkillRepository, additionalSkillProgressRepository);
   }
 
   @Bean
