+++ src/main/java/fr/avenirsesr/portfolio/student/progress/domain/model/AdditionalSkillProgress.java
@@ -1,5 +1,6 @@
-package fr.avenirsesr.portfolio.additionalskill.domain.model;
+package fr.avenirsesr.portfolio.student.progress.domain.model;
 
+import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkill;
 import fr.avenirsesr.portfolio.additionalskill.domain.model.enums.EAdditionalSkillLevel;
 import fr.avenirsesr.portfolio.common.data.domain.model.AvenirsBaseModel;
 import fr.avenirsesr.portfolio.user.domain.model.Student;
