+++ src/main/java/fr/avenirsesr/portfolio/student/progress/domain/service/StudentProgressServiceImpl.java
@@ -1,5 +1,11 @@
 package fr.avenirsesr.portfolio.student.progress.domain.service;
 
+import fr.avenirsesr.portfolio.additionalskill.domain.exception.AdditionalSkillNotFoundException;
+import fr.avenirsesr.portfolio.additionalskill.domain.exception.DuplicateAdditionalSkillException;
+import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkill;
+import fr.avenirsesr.portfolio.additionalskill.domain.model.enums.EAdditionalSkillLevel;
+import fr.avenirsesr.portfolio.additionalskill.domain.model.enums.EAdditionalSkillType;
+import fr.avenirsesr.portfolio.additionalskill.domain.port.output.repository.AdditionalSkillRepository;
 import fr.avenirsesr.portfolio.common.data.domain.model.PageCriteria;
 import fr.avenirsesr.portfolio.common.data.domain.model.PageInfo;
 import fr.avenirsesr.portfolio.common.data.domain.model.PagedResult;
@@ -7,9 +13,11 @@ import fr.avenirsesr.portfolio.common.data.domain.model.SortCriteria;
 import fr.avenirsesr.portfolio.program.domain.model.Skill;
 import fr.avenirsesr.portfolio.student.progress.domain.data.SkillLevelProgressWithTraceCountData;
 import fr.avenirsesr.portfolio.student.progress.domain.data.SkillProgressData;
+import fr.avenirsesr.portfolio.student.progress.domain.model.AdditionalSkillProgress;
 import fr.avenirsesr.portfolio.student.progress.domain.model.SkillLevelProgress;
 import fr.avenirsesr.portfolio.student.progress.domain.model.StudentProgress;
 import fr.avenirsesr.portfolio.student.progress.domain.port.input.StudentProgressService;
+import fr.avenirsesr.portfolio.student.progress.domain.port.output.repository.AdditionalSkillProgressRepository;
 import fr.avenirsesr.portfolio.student.progress.domain.port.output.repository.SkillLevelProgressRepository;
 import fr.avenirsesr.portfolio.student.progress.domain.port.output.repository.StudentProgressRepository;
 import fr.avenirsesr.portfolio.trace.domain.port.output.repository.TraceRepository;
@@ -28,6 +36,8 @@ public class StudentProgressServiceImpl implements StudentProgressService {
   private final StudentProgressRepository studentProgressRepository;
   private final SkillLevelProgressRepository skillLevelProgressRepository;
   private final TraceRepository traceRepository;
+  private final AdditionalSkillRepository additionalSkillRepository;
+  private final AdditionalSkillProgressRepository additionalSkillProgressRepository;
 
   @Override
   public boolean isStudentFollowingAPCProgram(Student student) {
@@ -116,7 +126,7 @@ public class StudentProgressServiceImpl implements StudentProgressService {
   }
 
   @Override
-  public PagedResult<SkillLevelProgress> search(
+  public PagedResult<SkillLevelProgress> searchSkillLevel(
       Student student, String keyword, PageCriteria pageCriteria) {
     log.debug("Searching SkillLevelProgress for {} with pagination {}", student, pageCriteria);
 
@@ -137,5 +147,44 @@ public class StudentProgressServiceImpl implements StudentProgressService {
                     .map(skillLevelProgress -> skillLevelProgress.getSkillLevel().getSkill()))
         .distinct()
         .toList();
+  }
+
+  @Override
+  public PagedResult<AdditionalSkillProgress> getAdditionalSkillsProgresses(
+      Student student, PageCriteria pageCriteria) {
+    return additionalSkillProgressRepository.findAllByStudent(student, pageCriteria);
+  }
+
+  @Override
+  public AdditionalSkillProgress createAdditionalSkillProgress(
+      Student student,
+      UUID additionalSkillId,
+      EAdditionalSkillType type,
+      EAdditionalSkillLevel level) {
+    try {
+      Optional<AdditionalSkill> additionalSkill =
+          additionalSkillRepository.findById(additionalSkillId);
+      AdditionalSkillProgress additionalSkillProgress =
+          AdditionalSkillProgress.create(
+              student, additionalSkill.orElseThrow(AdditionalSkillNotFoundException::new), level);
+      if (additionalSkillProgressRepository.additionalSkillProgressAlreadyExists(
+          additionalSkillProgress)) {
+        log.error(
+            "Failed to add additional skill [{}] for student [{}] because it already exists",
+            additionalSkillId,
+            student);
+        throw new DuplicateAdditionalSkillException();
+      }
+      return additionalSkillProgressRepository.save(additionalSkillProgress);
+    } catch (AdditionalSkillNotFoundException e) {
+      log.error("Failed to add additional skill for student [{}]: {}", student, e.getMessage());
+      throw e;
+    }
+  }
+
+  @Override
+  public PagedResult<AdditionalSkillProgress> searchAdditionalSkill(
+      Student student, String keyword, PageCriteria pageCriteria) {
+    return additionalSkillProgressRepository.findAllByStudent(student, pageCriteria, keyword);
   }
 }
