+++ src/main/java/fr/avenirsesr/portfolio/student/progress/infrastructure/adapter/repository/AdditionalSkillProgressJpaRepository.java
@@ -1,6 +1,6 @@
-package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.repository;
+package fr.avenirsesr.portfolio.student.progress.infrastructure.adapter.repository;
 
-import fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.model.AdditionalSkillProgressEntity;
+import fr.avenirsesr.portfolio.student.progress.infrastructure.adapter.model.AdditionalSkillProgressEntity;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
