+++ src/main/java/fr/avenirsesr/portfolio/student/progress/infrastructure/adapter/seeder/fake/FakeAdditionalSkillProgress.java
@@ -1,10 +1,10 @@
-package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.seeder.fake;
+package fr.avenirsesr.portfolio.student.progress.infrastructure.adapter.seeder.fake;
 
 import fr.avenirsesr.portfolio.additionalskill.domain.model.enums.EAdditionalSkillLevel;
 import fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.model.AdditionalSkillEntity;
-import fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.model.AdditionalSkillProgressEntity;
 import fr.avenirsesr.portfolio.common.seeder.domain.port.output.SharedDataGenerator;
 import fr.avenirsesr.portfolio.common.seeder.infrastructure.adapter.data.DataGeneratorProvider;
+import fr.avenirsesr.portfolio.student.progress.infrastructure.adapter.model.AdditionalSkillProgressEntity;
 import fr.avenirsesr.portfolio.user.infrastructure.adapter.model.StudentEntity;
 import java.util.List;
 import java.util.UUID;
