+++ src/main/java/fr/avenirsesr/portfolio/trace/application/adapter/mapper/TraceAssociationSearchResultMapper.java
@@ -1,7 +1,7 @@
 package fr.avenirsesr.portfolio.trace.application.adapter.mapper;
 
-import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkillProgress;
 import fr.avenirsesr.portfolio.ams.domain.model.AMS;
+import fr.avenirsesr.portfolio.student.progress.domain.model.AdditionalSkillProgress;
 import fr.avenirsesr.portfolio.student.progress.domain.model.SkillLevelProgress;
 import fr.avenirsesr.portfolio.trace.application.adapter.dto.ETraceAssociationType;
 import fr.avenirsesr.portfolio.trace.application.adapter.response.TraceAssociationSearchResult;
