+++ src/test/java/fr/avenirsesr/portfolio/student/progress/domain/service/StudentProgressServiceImplTest.java
@@ -1,10 +1,16 @@
 package fr.avenirsesr.portfolio.student.progress.domain.service;
 
+import static java.util.UUID.randomUUID;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
+import fr.avenirsesr.portfolio.additionalskill.domain.exception.AdditionalSkillNotFoundException;
+import fr.avenirsesr.portfolio.additionalskill.domain.exception.DuplicateAdditionalSkillException;
+import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkill;
+import fr.avenirsesr.portfolio.additionalskill.domain.model.enums.EAdditionalSkillLevel;
+import fr.avenirsesr.portfolio.additionalskill.domain.model.enums.EAdditionalSkillType;
+import fr.avenirsesr.portfolio.additionalskill.domain.port.output.repository.AdditionalSkillRepository;
 import fr.avenirsesr.portfolio.common.data.domain.model.PageCriteria;
 import fr.avenirsesr.portfolio.common.data.domain.model.PagedResult;
 import fr.avenirsesr.portfolio.common.data.domain.model.SortCriteria;
@@ -18,6 +24,7 @@ import fr.avenirsesr.portfolio.program.infrastructure.fixture.*;
 import fr.avenirsesr.portfolio.student.progress.domain.data.SkillLevelProgressWithTraceCountData;
 import fr.avenirsesr.portfolio.student.progress.domain.data.SkillProgressData;
 import fr.avenirsesr.portfolio.student.progress.domain.model.*;
+import fr.avenirsesr.portfolio.student.progress.domain.port.output.repository.AdditionalSkillProgressRepository;
 import fr.avenirsesr.portfolio.student.progress.domain.port.output.repository.StudentProgressRepository;
 import fr.avenirsesr.portfolio.student.progress.infrastructure.fixture.*;
 import fr.avenirsesr.portfolio.trace.domain.port.output.repository.TraceRepository;
@@ -40,6 +47,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 public class StudentProgressServiceImplTest {
   @Mock private StudentProgressRepository studentProgressRepository;
   @Mock private TraceRepository traceRepository;
+  @Mock private AdditionalSkillRepository additionalSkillRepository;
+  @Mock private AdditionalSkillProgressRepository additionalSkillProgressRepository;
   @InjectMocks private StudentProgressServiceImpl studentProgressService;
   private Student student;
 
@@ -839,6 +848,106 @@ public class StudentProgressServiceImplTest {
         BddLogger.then("it should return an empty list when no student progress is found");
         assertTrue(result.isEmpty(), "Result should be empty");
         verify(studentProgressRepository).findAllByStudent(eq(student));
+      }
+    }
+
+    @Nested
+    class WhenGettingAdditionalSkillsProgresses {
+
+      @Test
+      void getAdditionalSkillsProgresses_shouldDelegateToRepositoryAndReturnResult() {
+        BddLogger.given("the method getAdditionalSkillsProgresses");
+        Student localStudent = mock(Student.class);
+        PageCriteria criteria = new PageCriteria(1, 8);
+        PagedResult<AdditionalSkillProgress> expected = mock(PagedResult.class);
+
+        BddLogger.when("calling the method with a given student");
+        when(additionalSkillProgressRepository.findAllByStudent(localStudent, criteria))
+            .thenReturn(expected);
+
+        PagedResult<AdditionalSkillProgress> result =
+            studentProgressService.getAdditionalSkillsProgresses(localStudent, criteria);
+
+        BddLogger.then(
+            "it should return the expected paged additional skill progress and delegate to"
+                + " repository");
+        org.assertj.core.api.Assertions.assertThat(result).isSameAs(expected);
+        verify(additionalSkillProgressRepository).findAllByStudent(localStudent, criteria);
+      }
+    }
+
+    @Nested
+    class WhenCreatingAdditionalSkillProgress {
+      @Test
+      void createAdditionalSkillProgress_shouldSaveWhenSkillIsAvailableAndNotDuplicate() {
+        BddLogger.given("the method createAdditionalSkillProgress");
+        Student localStudent = mock(Student.class);
+        UUID skillId = randomUUID();
+        EAdditionalSkillType type = EAdditionalSkillType.ROME4;
+        EAdditionalSkillLevel level = EAdditionalSkillLevel.BEGINNER;
+        AdditionalSkill additionalSkill = mock(AdditionalSkill.class);
+
+        BddLogger.when("calling the method with an available and not duplicate skill");
+        when(additionalSkillRepository.findById(skillId)).thenReturn(Optional.of(additionalSkill));
+        when(additionalSkillProgressRepository.additionalSkillProgressAlreadyExists(any()))
+            .thenReturn(false);
+
+        studentProgressService.createAdditionalSkillProgress(localStudent, skillId, type, level);
+
+        BddLogger.then("it should save the additional skill progress");
+        verify(additionalSkillRepository).findById(skillId);
+        verify(additionalSkillProgressRepository).additionalSkillProgressAlreadyExists(any());
+        verify(additionalSkillProgressRepository).save(any(AdditionalSkillProgress.class));
+      }
+
+      @Test
+      void createAdditionalSkillProgress_shouldThrowDuplicateWhenAlreadyExists() {
+        BddLogger.given("the method createAdditionalSkillProgress");
+        Student localStudent = mock(Student.class);
+        UUID skillId = randomUUID();
+        EAdditionalSkillType type = EAdditionalSkillType.ROME4;
+        EAdditionalSkillLevel level = EAdditionalSkillLevel.BEGINNER;
+        AdditionalSkill additionalSkill = mock(AdditionalSkill.class);
+
+        BddLogger.when("calling the method with a duplicate skill");
+        when(additionalSkillRepository.findById(skillId)).thenReturn(Optional.of(additionalSkill));
+        when(additionalSkillProgressRepository.additionalSkillProgressAlreadyExists(any()))
+            .thenReturn(true);
+
+        BddLogger.then(
+            "it should throw a DuplicateAdditionalSkillException and not save the progress");
+        assertThrows(
+            DuplicateAdditionalSkillException.class,
+            () ->
+                studentProgressService.createAdditionalSkillProgress(
+                    localStudent, skillId, type, level));
+
+        verify(additionalSkillRepository).findById(skillId);
+        verify(additionalSkillProgressRepository).additionalSkillProgressAlreadyExists(any());
+        verify(additionalSkillProgressRepository, never()).save(any());
+      }
+
+      @Test
+      void createAdditionalSkillProgress_shouldRethrowWhenSkillNotFound() {
+        BddLogger.given("the method createAdditionalSkillProgress");
+        Student localStudent = mock(Student.class);
+        UUID skillId = randomUUID();
+        EAdditionalSkillType type = EAdditionalSkillType.ROME4;
+        EAdditionalSkillLevel level = EAdditionalSkillLevel.BEGINNER;
+
+        BddLogger.when("calling the method with an unknown skill");
+        when(additionalSkillRepository.findById(skillId))
+            .thenThrow(new AdditionalSkillNotFoundException());
+
+        BddLogger.then("it should throw an AdditionalSkillNotFoundException");
+        assertThrows(
+            AdditionalSkillNotFoundException.class,
+            () ->
+                studentProgressService.createAdditionalSkillProgress(
+                    localStudent, skillId, type, level));
+
+        verify(additionalSkillRepository).findById(skillId);
+        verifyNoInteractions(additionalSkillProgressRepository);
       }
     }
   }
